Update Url resolving
All data passed into a Url is now resolved into the Id.

This is so that we can include an Entity Id as a part of our Global Id and still have access to it later.

For those Urls in which there is no Id, just the Url is returned without any issue.